fix(mikroorm-types): changed local date property type (#238)
* fix(mikroorm-types): changed local date property type

* fix(mikroorm-types): updated test

+++ packages/mikroorm-types/src/LocalDateProperty.ts
@@ -21,6 +21,6 @@ export class LocalDateProperty extends Type {
 	}
 
 	getColumnType() {
-		return 'bigint';
+		return 'int';
 	}
 }
